Move the `memory_accessed` hook onto the `Extra` value
